layer.conf: BBMASK meta-ros fuse recipes
Using meta-ros/meta-ros2-jazzy solves the build issue where, without
any other changes but basecamp-image-everything will succeed when the
meta-ros/meta-ros2-jazzy layer is not in the build and fail when it is.
So BBMASK meta-ros fuse recipes.

ERROR: Nothing PROVIDES 'rviz-rendering' (but /scratch/sandeep/yocto/2025.1/yp-bc/sources/poky/../meta-ros/meta-ros2-jazzy/generated-recipes/fuse/fuse-viz_1.0.1-4.bb DEPENDS on or otherwise requires it)
rviz-rendering was skipped: Recipe will be skipped because: qt5: depends on qtbase; opengl: depends on rviz-ogre-vendor which depends on mesa which is not available because of missing opengl or vulkan in DISTRO_FEATURES; x11: depends on rviz-rendering which depends on rviz-ogre-vendor which depends on libx11,libxrandr,libxaw which require x11 in DISTRO_FEATURES

Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>