base-files: Add sbin to the users default path
Add sbin to the user's default path.  In most cases you need to be root (or
at least have some sort of elevated capabilities) to use most items in sbin,
so place these after the regular bin entries.

This is added for a better user experience when using sudo or similar
systems that elevate a users capabilities while keeping the environment.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>