basecamp-default-SM.wks: add to shrink unused partitions
Shrink the currently unused partitions to:
- help the image fit into various random 32GB-sized SDcards
- make the flashing process faster

The wic image is sized to 29GB and tested with the 32GB SDcard that I have.
Testing by other team members shows that even at this size it does not fit
onto all 32GB SDcards. Until we get more testing and can reach a better
consensus, and to make flashing faster in the mean time, shrink the size of
the wic image by slashing the sizes of the currently unused partitions.

Introduce a new wks file for the initial release but keep the previous file
for further investigation.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>