basecamp-disk-image: wic will not update the fstab with multiple rootfs
Instead of relying on WIC to update the disk image to match the partitions,
we do it manually by copying in a hand crafted fstab to the appropriate
rootfs during do_rootfs time.

Unfortunately do_fetch is unavailable in an image recipe, so we need to copy
the fstab(s) from a known location.  We do this file the base of the layer
similar to how the 'wic' directory is accessed.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>