xilinx-bootbin: Enable freertos only in default mc
The default multiconfig is where we build our overall images (ospi or disk).
We want to limit the freertos components to only being loaded when in this
configuration, otherwise it can try to add it into the image-recovery
partition.

Note, this might be better implemented as a negative of 'don't do this
in image recovery', as we may want to enable different multiconfigs
that DO have this behavior.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>