amd-edf: Clean up DISTRO_FEATURES
optee as a DISTRO_FEATURE is only used in this layer.  Remove to avoid
changing DISTRO_FEATURES by adding 'optee' to it.  Move the couple of
usages to MACHINE_FEATURES which is still being set.

The remaining two MACHINE_FEATURES based DISTRO_FEATURE settings are
libmali based and will still be set, while there should be another way
to handle this, we've sanitized all of the recipes that would link to
libmali to be MACHINE specific, avoiding the same sort of problem.

Further, add a vardepsexclude to the DISTRO_FEATURES so that any change
in MACHINE_FEATURES will NOT affect the resulting hashes.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>