systemd-conf: Add custom timesyncd conf
Add a custom timesyncd.conf drop-in that adds the internal Xilinx
ntp server in addition to the public google ntp server. The google
server needs to be listed explicitly in the conf file because if
the NTP option is set, it will override the built-in options, it
does not append. Therefore, both the public as well as the internal
ntp servers need to be listed for both scenarios to work.
time.google.com is the built-in default so we simply spell it out
in the conf file.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>