amd-edf: Update the way the version string is computed
A new 'minor' version is being introduced as a third digit. (default
to blank, for initial release.)

The build type is now relegated to development or release, as updates
will be captured in the minor version.

This will ensure that the repository URL can be based on the stable
version variable.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>