edf-disk-single-rootfs-efi.wks: update for mbr and 4k disk
With the first partition to --mbr; this requires an update to the --align
otherwise wic gives the following error:

	parted -s .../tmp-wic-ufs/edf-disk-single-rootfs-202507141821-sda.direct.mbr unit s mkpart primary fat32 1 34
	| output: Error: You requested a partition from 1s to 34s (sectors 1..34).
	| The closest location we can manage is 1s to 33s (sectors 1..33).

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>

Sync with John's patch, so this is only adjusting the alignment.

Also need to adjust the Boot0001 for efi due to the alignment changing.

We also change the alignment of the non-efi partition to match

Signed-off-by: Mark Hatle <mark.hatle@amd.com>