edf-linux-disk-image: Create UEFI image
Update edf-linux-disk-image and created a new efi specific
edf-disk-single-rootfs-efi to be a bootable UEFI image containing an
EFI System Partition (ESP), Kernel image and ubootefi.var persistent
environment file.

While this should not be necessary on non-EFI boot systems, the presence
of the ubootefi.vars and/or EFI/BOOT directories seems to convince
u-boot to always try EFI, even when specifically told to use distro boot.

This change allow the system to boot using a UEFI boot command and
allow the use of UFEI capsule updates. This also means that the system
can be booted using the `bootefi bootmgr` command rather than distro
boot (boot.scr).

Signed-off-by: John Toomey <john.toomey@amd.com>

Reformat the wks file for consistent spacing and order
Add missing --fstype=vfat to /boot/efi
Change the lable on all efi partitions to 'esp' and 'ESP' to match

Signed-off-by: Mark Hatle <mark.hatle@amd.com>