uefi-capsule: Fix race condition where boot.bin may be missing
The boot.bin is not deplayed into the deploy_dir until the 'do_deploy' task
is executed.  However, do_deploy is not the target of a recipe DEPENDS, so
you must explicitly depend on the task to ensure the file has been written.

Without this there is a small window where you can get an error that the
boot.bin file is not found.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>