u-boot-xlnx: Fix u-boot-initial-env generation
There were two issues observed.  The size of the environment is required
to generate the CRC correctly, but was hard coded to 0x4200, not the
value that u-boot was configured with.  This is configured with
CONFIG_ENV_SIZE.

For EDF ospi/qspi images, we have multiple u-boot variable stores, so the
'-r' option must be used to indicate we have redundant copies of the
environment in the spi.  This is configured by
CONFIG_SYS_REDUNDAND_ENVIRONMENT.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>