add toolchain to arm images
Add enough toolchain components to arm32 and aarch64 images such that the GNU
Hello sources can be unpacked, configured, built, installed, and run on the
target. Avoid the RISC-V targets since they use a ramdisk and their images
need to be kept lean.

For the edf-linux-disk-image, for example, the *wic.ufs size before and after
stays the same since it is a fixed size (6GB). Internally the amount of root
filesystem being used increases from 844,564 1k-blocks (15%) to 1,077,288
1k-blocks (19%).

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>