default-sc/local.conf.sample: Set AMD-EDF_VERSION in local.conf
The AMD-EDF_VERSION variable is set in the distro conf file but
for EDF OSPI builds that include the Linux image recovery, the
variable is not set as we are using a different distro for it.

The variable is also used in the sstate and source mirror paths
specified in local.conf. Therefore, explicitely set the variable
in local.conf so it is set no matter which distro is used.
Otherwise, the Linux image recovery build will not be able to use
sstate or dl mirrors.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>

Copy of previous commit but for the new default-sc template.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>