*capsule-install: Remove recipes
Remove the vek280 and vek385 capsule install recipes. That's not
the right way for deploying and installing system firmware updates.
System firmware updates should be uploaded to lvfs so a user can
install updates from there using fwupdmgr.

There are several pieces missing to enable that flow end-to-end
but we should not promote this method as a workaround nor expand
it to other platforms.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>