xen: Split xen-efi-config file for each soc
Some of the xen configurations values varies for each soc it is better
to split xen-efi-config file for each soc(versal and versal-2ve-2vm).

Also bump dom0_max_vcpus value for 1 to 6 so that we have more number
of vcpus allocated to dom0 which allows to boot additional vm's after
dom0 while stil keeping the null scheduler.

Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Message-ID: <20250929173858.1362359-1-sandeep.gundlupet-raju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>