amd-edf: Enable rauc update tool and related systems
Note, this does not include the rauc components into the filesystem, but
does enable the DISTRO_FEATURE and availability of the components to be
added to a filesystem if desired.

curl change has been bbmasked out, as this is an enhancement that we do
not need.  It has the side effect of causing elfutils to rebuild, and many
system components depend on elfutils.  We can remove this mask in the
future if the behavior change is desired.

packagegroup-base change has been bbmasked out, as this change will force
all images to include rauc, even if the image does not plan to use it.
Instead, the image is now responsible for including the 'rauc' package.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-Id: <20251021205803.1111868-2-mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>