bootbin-version-header.bbappend: Write rollback counter as binary instead of UTF-8 text
The rollback counter file was previously written using .decode('utf-8')
on the integer-to-bytes conversion, which is incorrect since the data is
binary and not valid UTF-8 text.

Updated the logic to write the EDF version in 4-byte little-endian
binary format using 'wb' mode. This ensures the rollback counter file
contains the correct binary representation.

Zynqmp doesnot use bootgen optionaldata field support so it still uses
the hex.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Message-ID: <20251022101327.3078857-1-rajukumar.pothuraju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>