amd-edf: Enable rauc configurations by default
The rauc distro feature is not enough to enable a few specific configurations.

We disable (BBMASK) the u-boot-xlnx-scr change, as this will be done in a
different way that is system-controller specific.

The remaining items, system controller specific packages and the linux-xlnx
recipe will now be generated with the rauc components enabled by default.

In the future we can likely rework this code to simply use the
DISTRO_FEATURE, but set these switches until then.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-Id: <20251023223032.1600029-1-mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>