edf-linux-disk-image: Do not limit rootfs UUID by SoC
The ROOTFS UUID is not used in all images, but there is no reason to not
generate it.  Also the way it is generated ensures that regular (512 byte)
and UFS (4k) will get different UUIDs, hopefully avoiding confusion if a
user tries to have both filesystems on the same machine.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-ID: <20251113202248.1813262-3-mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>