amd-edf.conf: Move partition GUIDs
Move the pre-defined partition type GUIDs to the EDF distro
configuration, rename to reflect the specific architecture and add other
architecture specific GUIDs for future use.

Move the WICVARS and partition UUID generation code into a class so that
it can be used by multiple images and pull it into the
edf-linux-disk-image recipe.

Signed-off-by: John Toomey <john.toomey@amd.com>
Message-ID: <20251118140412.3114319-2-john.toomey@amd.com>

Minor rework to use overrides for ROOTFS_PART_TYPE.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>